move separate classes into their own files for clarity
